sorted the images by placing them in one location to ease the import function for images
