+++ front_end/src/App.jsx
@@ -1,4 +1,4 @@
-import Homepage from "./components/Homepage"
+import Homepage from "./containers/Homepage"
 
 function App() {
   return (
