+++ front_end/src/assets/entrance02.png
