+++ front_end/src/assets/hero-section.png
