+++ front_end/src/assets/standard_room.png
