+++ front_end/src/assets/superior_room.png
