+++ front_end/src/assets/sweeties-restaurant.png
