--- front_end/src/components/Homepage/Footer/footer.scss
@@ -1,62 +1,0 @@
-footer{
-    background: #000;
-    color: #fff;    
-    .footer_wrapper{
-        padding: 10px 12px;
-        h4{
-            text-align: center;
-            font-size: 1.2em;
-            font-weight: bold;
-        }
-        .socials{
-            ul{
-                list-style: none;
-                li{
-                    margin: 0.5em 0px;
-                    letter-spacing: 0.5px;
-                    a{
-                        text-decoration: none;
-                        color: #fff;
-                        .icon-instagram{
-                            background: linear-gradient(45deg, #405de6, #5851db, #833ab4, #c13584, #e1306c, #fd1d1d);
-                        }
-                    }
-                }
-            }            
-        }   
-        .location{
-            iframe{
-                width: 100%;
-                height: 400px;
-                margin: auto;
-            }
-        }     
-    }
-    .copyright{
-        border-top: 1px solid #868686;
-        padding: 4px 0px;
-        p{
-            margin: unset;
-            text-align: center;
-        }
-    }
-}
-
-
-@media screen and (width >= 726px) {
-    footer{        
-        .footer_wrapper{
-            display: grid;
-            grid-template-columns: 1fr 1fr;            
-            padding: 10px 2%;
-            h4{
-                text-align: left;
-            }
-            .location{
-                iframe{
-                    height: 250px;
-                }
-            }
-        }
-    }
-}
