+++ front_end/src/constants/images/Images.jsx
@@ -1,0 +1,33 @@
+import hero from '../../assets/hero-section.png'
+import big5_lounge from '../../assets/big5-lounge.png'
+import lovers_garden from '../../assets/lovers-garden.png'
+import sweeties_restaurant from '../../assets/sweeties-restaurant.png'
+import logo from '../../assets/logo.png'
+import entrance01 from '../../assets/entrance01.png'
+import entrance02 from '../../assets/entrance02.png'
+import screen from '../../assets/screen.png'
+import lovers from '../../assets/lovers_garden.png'
+import wine_glass from '../../assets/wine_glass.png'
+import standard_room from '../../assets/standard_room.png'
+import superior_room from '../../assets/superior_room.png'
+
+
+export default function Images() {
+    return(
+        hero, 
+        big5_lounge, 
+        lovers_garden, 
+        sweeties_restaurant,
+        logo,
+        entrance01,
+        entrance02,
+        screen,
+        lovers,
+        wine_glass,
+        standard_room,
+        superior_room
+    )
+}
+
+
+
