+++ front_end/src/containers/Homepage/About/about.css
@@ -18,6 +18,7 @@
   height: 400px;
   margin: 1em 0px;
   position: relative;
+  background: linear-gradient(rgb(130, 1, 180), rgb(129, 128, 128));
 }
 .About .our_story .story img {
   height: 100%;
