+++ front_end/src/containers/Homepage/About/about.scss
@@ -21,6 +21,7 @@
             height: 400px;
             margin: 1em 0px;
             position: relative;
+            background: linear-gradient(rgb(130, 1, 180), rgb(129, 128, 128));
             img{
                 height: 100%;
                 width: 100%;
