+++ front_end/src/containers/Homepage/Footer/footer.css
@@ -1,0 +1,99 @@
+footer {
+  background: #000;
+  color: #fff;
+}
+footer .footer_wrapper {
+  padding: 10px 12px;
+  text-align: center;
+}
+footer .footer_wrapper h4 {
+  font-size: 1.2em;
+  font-weight: bold;
+  margin: 0px 0px 0.8em;
+}
+footer .footer_wrapper .logo,
+footer .footer_wrapper .socials,
+footer .footer_wrapper .location,
+footer .footer_wrapper .working_hours {
+  margin: 8px 0px 2.5em;
+}
+footer .footer_wrapper .logo img {
+  width: 100px;
+}
+footer .footer_wrapper .logo h4 {
+  margin: unset;
+}
+footer .footer_wrapper .logo p {
+  margin: 8px 0px;
+}
+footer .footer_wrapper .socials ul {
+  list-style: none;
+}
+footer .footer_wrapper .socials ul li {
+  margin: 0.5em 0px;
+  letter-spacing: 0.5px;
+}
+footer .footer_wrapper .socials ul li a {
+  text-decoration: none;
+  color: #fff;
+}
+footer .footer_wrapper .socials ul li a .icon-instagram {
+  background: linear-gradient(45deg, #405de6, #5851db, #833ab4, #c13584, #e1306c, #fd1d1d);
+}
+footer .footer_wrapper .location iframe {
+  width: 100%;
+  height: 400px;
+  margin: auto;
+}
+footer .footer_wrapper .working_hours .info p {
+  margin: 4px 0px;
+}
+footer .copyright {
+  border-top: 1px solid #424242;
+  padding: 16px 0px;
+}
+footer .copyright p {
+  margin: unset;
+  text-align: center;
+}
+
+@media screen and (width >= 726px) {
+  footer .footer_wrapper {
+    display: grid;
+    grid-template-columns: auto auto auto auto;
+    -moz-column-gap: 1em;
+         column-gap: 1em;
+    padding: 10px 2%;
+    text-align: left;
+  }
+  footer .footer_wrapper .logo,
+  footer .footer_wrapper .socials,
+  footer .footer_wrapper .working_hours,
+  footer .footer_wrapper .location {
+    padding: 8px 0px;
+    margin: 0px;
+  }
+  footer .footer_wrapper .logo h4,
+  footer .footer_wrapper .socials h4,
+  footer .footer_wrapper .working_hours h4,
+  footer .footer_wrapper .location h4 {
+    text-align: left;
+  }
+  footer .footer_wrapper .logo {
+    display: flex;
+    align-items: center;
+  }
+  footer .footer_wrapper .logo h4 {
+    font-size: 1.5em;
+  }
+  footer .footer_wrapper .logo p {
+    text-align: center;
+    font-style: italic;
+  }
+  footer .footer_wrapper .location iframe {
+    height: 200px;
+  }
+  footer .footer_wrapper .working_hours .info {
+    padding: 0px;
+  }
+}/*# sourceMappingURL=footer.css.map */
