+++ front_end/src/containers/Homepage/Footer/footer.scss
@@ -1,0 +1,112 @@
+footer{
+    background: #000;
+    color: #fff;    
+    .footer_wrapper{
+        padding: 10px 12px;
+        text-align: center;
+        h4{
+            font-size: 1.2em;
+            font-weight: bold;
+            margin: 0px 0px 0.8em;
+        }
+        .logo,
+        .socials,
+        .location,
+        .working_hours{
+            margin: 8px 0px 2.5em;
+        }
+        .logo{
+            img{
+                width: 100px;
+            }
+            h4{
+                margin: unset;
+            }
+            p{
+                margin: 8px 0px;
+            }
+        }
+        .socials{
+            ul{
+                list-style: none;
+                li{
+                    margin: 0.5em 0px;
+                    letter-spacing: 0.5px;
+                    a{
+                        text-decoration: none;
+                        color: #fff;
+                        .icon-instagram{
+                            background: linear-gradient(45deg, #405de6, #5851db, #833ab4, #c13584, #e1306c, #fd1d1d);
+                        }
+                    }
+                }
+            }            
+        }   
+        .location{
+            iframe{
+                width: 100%;
+                height: 400px;
+                margin: auto;
+            }
+        }     
+        .working_hours{
+            .info{
+                p{
+                    margin: 4px 0px;
+                }
+            }
+        }
+    }
+    .copyright{
+        border-top: 1px solid #424242;
+        padding: 16px 0px;
+        p{
+            margin: unset;
+            text-align: center;
+        }
+    }
+}
+
+
+@media screen and (width >= 726px) {
+    footer{        
+        .footer_wrapper{
+            display: grid;
+            grid-template-columns: auto auto auto auto;      
+            column-gap: 1em;      
+            padding: 10px 2%;
+            text-align: left;
+            .logo,
+            .socials,
+            .working_hours,
+            .location{
+                padding: 8px 0px;      
+                margin: 0px;                          
+                h4{
+                    text-align: left;                    
+                }
+            }                 
+            .logo{
+                display: flex;
+                align-items: center;
+                h4{
+                    font-size: 1.5em;
+                }
+                p{
+                    text-align: center;
+                    font-style: italic;
+                }
+            }       
+            .location{
+                iframe{
+                    height: 200px;
+                }
+            }
+            .working_hours{                
+                .info{                    
+                    padding: 0px;                    
+                }
+            }
+        }
+    }
+}
