+++ front_end/src/containers/Homepage/Gallery/gallery.scss
@@ -12,6 +12,7 @@
         }
         .gallery{
             margin: 1em 0px;
+            background: linear-gradient(rgb(3, 3, 3), rgb(129, 128, 128));
             img{
                 width: 100%;
                 height: 100%;
