+++ front_end/src/containers/Homepage/Hero/hero.css
@@ -3,7 +3,7 @@
   justify-content: center;
   width: 100%;
   height: 93vh;
-  background: url(./images/hero_section.png);
+  background: url(../../../assets/hero-section.png);
   background-position: center;
   background-size: cover;
   background-repeat: no-repeat;
