+++ front_end/src/containers/Homepage/Rooms/rooms.scss
@@ -11,6 +11,7 @@
             margin: 18px 0;
             .image{
                 height: 18em;
+                background: linear-gradient(rgb(3, 3, 3), rgb(129, 128, 128));
                 img{
                     width: 100%;
                     height: 100%;
